\docType{package}
\name{cdtRintroduction-package}
\alias{cdtRintroduction}
\alias{cdtRintroduction-package}
\alias{cdtrintroduction}
\title{R courses for Newcastle University's CDT in Cloud Computing for Big Data}
\description{
  Functions and datasets used for Newcastle University's 
  CDT in Cloud Computing for Big Data. This package is 
  used in the "Introduction to R" course.
}
\keyword{package}

